Fix non responsive touchscreen for ATMEL_X0080_X0026 touch controller
HTC recently started including another version of the touch controller
in incredibles. The mutli touch detection previously only looked at the
controller family to assign the correct number of MT points. Change it to
look at the family and controller version to assign the number of points.
If a controller that is not listed is found, fall back to the default values
defined in the board file.